Initial Commit
Currently consists of 2 moving obstacles with a set path and USV which can alter is course. Currently existing in an integer domain for simplicity, will likely be moved to real/continuous later. USV becomes damaged if it reaches within 5 units x and y of an obstacle. Similarly it reaches the goal if it is within 5 units of (1000,1000). 

Does not currently work with the JAX planner, but the Random Agent works just fine, so syntax seems correct. Seems to be an issue with the JAX planner itself.